Refactor to stop unnecessarily exposing public APIs
