+++ portfolio/src/test/java/com/google/sps/servlets/ServletTest.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.sps;
+package com.google.sps.servlets;
 
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
